+++ kaos_ui/__init__.py
@@ -18,7 +18,7 @@ from kaos_ui.manifest import (
     register_template,
     resolve_kind,
 )
-from kaos_ui.runtime import register_kaos_ui_tools
+from kaos_ui.runtime import register_kaos_ui_tools, register_ui_tools
 from kaos_ui.scaffolder import ScaffoldResult, scaffold
 from kaos_ui.settings import KaosUISettings
 
@@ -39,6 +39,7 @@ __all__ = [
     "register_alias",
     "register_kaos_ui_tools",
     "register_template",
+    "register_ui_tools",
     "resolve_kind",
     "scaffold",
 ]